cv02 zmeny
