+++ www/gulm04/cv02-ukazky/cv02/includes/footer.php
@@ -1,0 +1,3 @@
+</body>
+
+</html>
