+++ www/gulm04/cv02-ukazky/cv02/styles/style.css
@@ -1,0 +1,53 @@
+* {
+    margin: 0 auto;
+    padding: 0 auto;
+    font-family: 'Courier New', Courier, monospace;
+}
+body {
+    display: flex;
+    flex-direction: column;
+}
+h1 {
+    font-size: 2rem;
+    margin: 15px auto;
+    text-align: center;
+}
+h3 {
+    font-size: 1.5rem;
+    margin-bottom: 30px;
+}
+img {
+    width: 60px;
+    height: 50px;
+    padding: 10px;
+}
+.b-card {
+    display: flex;
+    width: 800px;
+}
+.container, .contact {
+    width: 300px;
+    height: 100px;
+    background-color: #81d4fa;
+    -webkit-box-shadow: 2px 3px 28px -1px rgba(0,0,0,0.57); 
+    box-shadow: 2px 3px 28px -1px rgba(0,0,0,0.57);
+    font-size: 1.2rem;
+    padding: 5px;
+    margin: 10px auto;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+}
+.container {
+    flex-direction: row;
+}
+.contact {
+    margin-bottom: 60px;
+}
+.personal {
+    padding: 10px;
+}
+.title {
+    font-weight: bold;
+}
